Rename constants to colors. Add pointer to all interactive elements
